chore(suite): split FormattedDate and FormattedDateWithBullet into two files
